Trying out live update of room lobby
